% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forestControl.R
\docType{package}
\name{forestControl-package}
\alias{forestControl-package}
\title{False Positive Rate Control in Selection Frequency for Random Forest}
\description{
This package is an implementation of the methods described by Ender Konukoglu and Melanie Ganz in
\emph{Konukoglu, E. and Ganz, M., 2014. Approximate false positive rate control in selection frequency
for random forest. arXiv preprint arXiv:1410.2838} \url{https://arxiv.org/abs/1410.2838}.
}
