% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpr_fs.R
\name{fpr_fs}
\alias{fpr_fs}
\title{False Postivie Rate Feature Selection}
\usage{
fpr_fs(x)
}
\arguments{
\item{x}{a \code{randomForest} or \code{ranger} object}
}
\value{
a \code{tibble} of selection frequencies and their false positive rate
}
\description{
Calculate the False Positive Rate (FPR) for each feature using it's selection frequency
}
\examples{
library(randomForest)
data(iris)
iris.rf <- randomForest(iris[,-5], iris[,5], forest = TRUE)

iris.features <- fpr_fs(iris.rf)
print(iris.features)
}
\author{
Jasen Finch \email{jsf9@aber.ac.uk}
}
