% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sft-method.R, R/plot_sft.R
\name{plot_sft}
\alias{plot_sft}
\alias{plot_sft.randomForest}
\alias{plot_sft.ranger}
\title{Plot Selection Frequency Threshold}
\usage{
plot_sft(x)

\method{plot_sft}{randomForest}(x)

\method{plot_sft}{ranger}(x)
}
\arguments{
\item{x}{a randomForest of ranger forest object}
}
\value{
a plot
}
\description{
Plot Selection Frequency Threshold
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

