% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecFreqs.R
\name{selecFreqs}
\alias{selecFreqs}
\title{Variable Selection Frequencies}
\usage{
selecFreqs(x)
}
\arguments{
\item{x}{a \code{randomForest} or \code{ranger} object}
}
\value{
\code{data.frame} of variable selection frequencies
}
\description{
Extract variable selection frequencies from forest objects
}
