% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sft_calc.R
\name{sft_calc}
\alias{sft_calc}
\title{Selection Frequency Threshold}
\usage{
sft_calc(Ft, Fn, K, Tr, alpha)
}
\arguments{
\item{Ft}{the total number of features}

\item{Fn}{the number of features considered at each internal node (mtry)}

\item{K}{the average number of binary tests/internal nodes across the enitre forest}

\item{Tr}{the total number of trees in the forest}

\item{alpha}{a false positive rate (ie, 0.01)}
}
\value{
a list of two elements
\describe{
    \item{sft}{the selection frequency threshold}
    \item{probs_atsft}{the esimated false positive rate}
}
}
\description{
Calculate the selection frequency threshold for a give approximate false postive rate
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}

Jasen Finch \email{jsf9@aber.ac.uk}
}
\seealso{
\code{\link{extract_params}}
}

