% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sftPlot.R
\name{sftPlot}
\alias{sftPlot}
\title{Plot Selection Frequency Threshold}
\usage{
sftPlot(x)
}
\arguments{
\item{x}{a randomForest of ranger forest object}
}
\value{
a plot
}
\description{
Plot Selection Frequency Threshold
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}

