% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sft.R
\name{sft}
\alias{sft}
\title{Selection Frequency Threshold}
\usage{
sft(x, alpha)
}
\arguments{
\item{x}{a \code{randomForest} or \code{ranger} object}

\item{alpha}{a false positive rate (ie, 0.01)}
}
\value{
a list of two elements
\describe{
    \item{sft}{the selection frequency threshold}
    \item{probs_atsft}{the esimated false positive rate}
}
}
\description{
Selection Frequency Threshold
}
\author{
Tom Wilson \email{tpw2@aber.ac.uk}
}
